Change the prototype of parsec_gpu_get_best_device
 - parsec_gpu_get_best_device now takes as a parameter which task is
   being scheduled, and returns both the device to use (return value)
   and an estimate of nanosecond of compute time that this would occupy
   said device.
 - the parsec_device_load array is not updated by the operation anymore,
   this update is left to the parsec_XXX_kernel_scheduler function, so
   that addition and substraction are done symmetrically in the same
   file (and the load is only accounted for while the task is actually
   inside a queue or scheduled to the device)
 - to compute time estimates, task classes have a time_estimate member
   that takes the pointer to the task and the pointer to the device.
   time_estimate returns an int64_t which is the number of nanoseconds
   estimated.
 - the device exposes some internal capacity (typically, the number
   of floating point operation of a given type per nanosecond)
 - PTG replaces the 'weight' BODY property (which becomes obsolete)
   with the 'time_estimate' TASK CLASS property. That property is
   the name of the time_estimate function to use.
 - load is now updated atomically by parsec_XXX_kernel_scheduler,
   using the value returned by get_best_device and then stored in
   the gpu task structure. Atomic updates ensure that the current
   load of devices is accurate, and that it returns to 0 when all
   tasks are done executing.

Transition PTG from weight-based load balancing to time heuristic

Transition DTD from weight-based load balancing to time heuristic

Port tests to time estimate heuristic using both methods available

+++ parsec/mca/device/device_gpu.h
@@ -93,7 +93,7 @@ struct parsec_gpu_task_s {
         struct {
             parsec_task_t           *ec;
             uint64_t                 last_data_check_epoch;
-            double                   load;  /* computational load imposed on the device */
+            uint64_t                 load;  /* computational load imposed on the device */
             const parsec_flow_t     *flow[MAX_PARAM_COUNT];
             uint32_t                 flow_nb_elts[MAX_PARAM_COUNT]; /* for each flow, size of the data to be allocated
                                                                      * on the GPU.
